requirements.txt
requests
pandas
sqlalchemy
mysql-connector-python
websocket-client

